+++ src/DevilDaggersInfo.Core.Mod/FileHandling/PngFileHandler.cs
@@ -1,3 +1,7 @@
+using SixLabors.ImageSharp;
+using SixLabors.ImageSharp.PixelFormats;
+using SixLabors.ImageSharp.Processing;
+
 namespace DevilDaggersInfo.Core.Mod.FileHandling;
 
 internal sealed class PngFileHandler : IFileHandler